+++ osu.Game.Tests/Visual/SongSelect/TestScenePlaySongSelect.cs
@@ -96,6 +96,7 @@ namespace osu.Game.Tests.Visual.SongSelect
 
             AddAssert("filter count is 1", () => songSelect.FilterCount == 1);
         }
+
         [Test]
         public void TestChangeBeatmapBeforeEnter()
         {
@@ -117,6 +118,7 @@ namespace osu.Game.Tests.Visual.SongSelect
                 InputManager.ReleaseKey(Key.Enter);
             });
 
+            AddUntilStep("wait for not current", () => !songSelect.IsCurrentScreen());
             AddAssert("ensure selection changed", () => selected != Beatmap.Value);
 
             AddUntilStep("wait for return to song select", () => songSelect.IsCurrentScreen());
@@ -144,6 +146,7 @@ namespace osu.Game.Tests.Visual.SongSelect
                 InputManager.ReleaseKey(Key.Down);
             });
 
+            AddUntilStep("wait for not current", () => !songSelect.IsCurrentScreen());
             AddAssert("ensure selection didn't change", () => selected == Beatmap.Value);
 
             AddUntilStep("wait for return to song select", () => songSelect.IsCurrentScreen());
@@ -175,6 +178,7 @@ namespace osu.Game.Tests.Visual.SongSelect
                 InputManager.ReleaseKey(Key.Enter);
             });
 
+            AddUntilStep("wait for not current", () => !songSelect.IsCurrentScreen());
             AddAssert("ensure selection changed", () => selected != Beatmap.Value);
 
             AddUntilStep("wait for return to song select", () => songSelect.IsCurrentScreen());
@@ -207,6 +211,7 @@ namespace osu.Game.Tests.Visual.SongSelect
                 InputManager.ReleaseButton(MouseButton.Left);
             });
 
+            AddUntilStep("wait for not current", () => !songSelect.IsCurrentScreen());
             AddAssert("ensure selection didn't change", () => selected == Beatmap.Value);
 
             AddUntilStep("wait for return to song select", () => songSelect.IsCurrentScreen());