+++ osu.Game/Screens/Edit/Compose/Components/SelectionHandler.cs
@@ -238,16 +238,16 @@ namespace osu.Game.Screens.Edit.Compose.Components
 
                 var items = new List<MenuItem>
                 {
-                    new OsuMenuItem("Sound")
+                    new OsuMenuItem("音效")
                     {
                         Items = new[]
                         {
-                            createHitSampleMenuItem("Whistle", HitSampleInfo.HIT_WHISTLE),
-                            createHitSampleMenuItem("Clap", HitSampleInfo.HIT_CLAP),
-                            createHitSampleMenuItem("Finish", HitSampleInfo.HIT_FINISH)
+                            createHitSampleMenuItem("口哨", HitSampleInfo.HIT_WHISTLE),
+                            createHitSampleMenuItem("拍掌", HitSampleInfo.HIT_CLAP),
+                            createHitSampleMenuItem("重音", HitSampleInfo.HIT_FINISH)
                         }
                     },
-                    new OsuMenuItem("Delete", MenuItemType.Destructive, deleteSelected),
+                    new OsuMenuItem("删除", MenuItemType.Destructive, deleteSelected),
                 };
 
                 if (selectedBlueprints.Count == 1)