+++ osu.Game/Screens/Play/PlayerSettings/VisualSettings.cs
@@ -27,12 +27,18 @@ namespace osu.Game.Screens.Play.PlayerSettings
                 {
                     Text = "背景暗化:"
                 },
-                dimSliderBar = new PlayerSliderBar<double>(),
+                dimSliderBar = new PlayerSliderBar<double>
+                {
+                    DisplayAsPercentage = true
+                },
                 new OsuSpriteText
                 {
                     Text = "背景模糊:"
                 },
-                blurSliderBar = new PlayerSliderBar<double>(),
+                blurSliderBar = new PlayerSliderBar<double>
+                {
+                    DisplayAsPercentage = true
+                },
                 new OsuSpriteText
                 {
                     Text = "切换:"